Fix : core/index.js파일 분리, router.js, store.js 추가
