+++ src/store/myCurrentValley.js
@@ -1,4 +1,4 @@
-import { Store } from '../core/index.js';
+import { Store } from '../core/store.js';
 
 //image store
 const store = new Store({
